Improve transition from previous behavior
- Simplify Model\User methods parameter list
- Use DBA methods in mod/regmod
- Replace killme with exit in mod/regmod
- Simplify goaway() calls

+++ mod/regmod.php
@@ -20,44 +20,34 @@ function user_allow($hash)
 	$a = get_app();
 
 	$register = Register::getByHash($hash);
-
 	if (!DBA::isResult($register)) {
 		return false;
 	}
 
-	$user = q("SELECT * FROM `user` WHERE `uid` = %d LIMIT 1",
-		intval($register['uid'])
-	);
-
+	$user = User::getById($register['uid']);
 	if (!DBA::isResult($user)) {
-		killme();
+		exit();
 	}
 
 	Register::deleteByHash($hash);
 
-	$r = q("UPDATE `user` SET `blocked` = 0, `verified` = 1 WHERE `uid` = %d",
-		intval($register['uid'])
-	);
+	DBA::update('user', ['blocked' => false, 'verified' => true], ['uid' => $register['uid']]);
 
-	$r = q("SELECT * FROM `profile` WHERE `uid` = %d AND `is-default` = 1",
-		intval($user[0]['uid'])
-	);
-	if (DBA::isResult($r) && $r[0]['net-publish']) {
-		$url = System::baseUrl() . '/profile/' . $user[0]['nickname'];
-		if ($url && strlen(Config::get('system', 'directory'))) {
-			Worker::add(PRIORITY_LOW, "Directory", $url);
-		}
+	$profile = DBA::selectFirst('profile', ['net-publish'], ['uid' => $register['uid'], 'is-default' => true]);
+
+	if (DBA::isResult($profile) && $profile['net-publish'] && Config::get('system', 'directory')) {
+		$url = System::baseUrl() . '/profile/' . $user['nickname'];
+		Worker::add(PRIORITY_LOW, "Directory", $url);
 	}
 
 	L10n::pushLang($register['language']);
 
 	$res = User::sendRegisterOpenEmail(
-		$user[0]['email'],
+		$user,
 		Config::get('config', 'sitename'),
-		System::baseUrl(),
-		$user[0]['username'],
-		'Sent in a previous email',
-		$user[0]);
+		$a->getBaseUrl(),
+		defaults($register, 'password', 'Sent in a previous email')
+	);
 
 	L10n::popLang();
 
@@ -73,20 +63,20 @@ function user_allow($hash)
 function user_deny($hash)
 {
 	$register = Register::getByHash($hash);
-
 	if (!DBA::isResult($register)) {
 		return false;
 	}
 
-	$user = q("SELECT * FROM `user` WHERE `uid` = %d LIMIT 1",
-		intval($register['uid'])
-	);
+	$user = User::getById($register['uid']);
+	if (!DBA::isResult($user)) {
+		exit();
+	}
 
 	DBA::delete('user', ['uid' => $register['uid']]);
 
 	Register::deleteByHash($register['hash']);
 
-	notice(L10n::t('Registration revoked for %s', $user[0]['username']) . EOL);
+	notice(L10n::t('Registration revoked for %s', $user['username']) . EOL);
 	return true;
 }
 
@@ -94,17 +84,16 @@ function regmod_content(App $a)
 {
 	if (!local_user()) {
 		info(L10n::t('Please login.') . EOL);
-		$o = '<br /><br />' . Login::form($a->query_string, intval(Config::get('config', 'register_policy')) === REGISTER_CLOSED ? 0 : 1);
-		return $o;
+		return Login::form($a->query_string, intval(Config::get('config', 'register_policy')) === REGISTER_CLOSED ? 0 : 1);
 	}
 
-	if ((!is_site_admin()) || (x($_SESSION, 'submanage') && intval($_SESSION['submanage']))) {
+	if (!is_site_admin() || !empty($_SESSION['submanage'])) {
 		notice(L10n::t('Permission denied.') . EOL);
 		return '';
 	}
 
 	if ($a->argc != 3) {
-		killme();
+		exit();
 	}
 
 	$cmd = $a->argv[1];
@@ -112,13 +101,11 @@ function regmod_content(App $a)
 
 	if ($cmd === 'deny') {
 		user_deny($hash);
-		goaway(System::baseUrl() . "/admin/users/");
-		killme();
+		goaway('admin/users/');
 	}
 
 	if ($cmd === 'allow') {
 		user_allow($hash);
-		goaway(System::baseUrl() . "/admin/users/");
-		killme();
+		goaway('admin/users/');
 	}
 }
